requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
backoff==2.2.1
bcrypt==5.0.0
build==1.4.2
certifi==2026.2.25
charset-normalizer==3.4.7
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.3.2
colorama==0.4.6
distro==1.9.0
dnspython==2.8.0
durationpy==0.10
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.24
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.3.0
googleapis-common-protos==1.74.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
huggingface_hub==1.9.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
kubernetes==35.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
numpy @ file:///D:/bld/numpy_1707225528937/work/dist/numpy-1.26.4-cp311-cp311-win_amd64.whl#sha256=537f0573fefb1e93b3633e8a79102e6b6d2911e09e0a90008993efef4242448e
oauthlib==3.3.1
onnxruntime==1.24.4
openai==1.30.5
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-instrumentation==0.61b0
opentelemetry-instrumentation-asgi==0.61b0
opentelemetry-instrumentation-fastapi==0.61b0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-util-http==0.61b0
orjson==3.11.8
overrides==7.7.0
packaging==26.0
posthog==7.9.12
protobuf==6.33.6
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pypdf==4.2.0
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.9
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
rich==14.3.3
rich-toolkit==0.19.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.14.0
tenacity==9.1.4
tiktoken==0.7.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.12.0
urllib3==2.6.3
uvicorn==0.29.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
wrapt==1.17.3
zipp==3.23.0
